feature_services

+++ src/main/java/ru/skypro/dto/SocksDto.java
@@ -6,14 +6,20 @@ import lombok.Getter;
 import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 
+/**
+ * DTO с информацией о носках
+ */
 @Getter
 @ToString
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SocksDto {
 
+    //цвет
     String color;
-    int cottonPart;
+    //процент хлопка
+    short cottonPart;
+    //количество
     int quantity;
 
 }
